04-02-23

+++ app/Http/Responses/LoginResponse.php
@@ -21,16 +21,26 @@ class LoginResponse implements LoginResponseContract
         $request->session()->regenerateToken();
         return redirect(route("login_guest"));*/
 
-        $request->session()->regenerate();
-        $previous_session = Auth::User()->session_id;
-        if ($previous_session) {
-            Session::getHandler()->destroy($previous_session);
+        if ($request->user()->status == 'activo') {
+
+            $request->session()->regenerate();
+            $previous_session = Auth::User()->session_id;
+            if ($previous_session) {
+                Session::getHandler()->destroy($previous_session);
+            }
+
+            $user= User::where('id',Auth::user()->id)->first();
+            $user->session_id = Session::getId();
+            $user->save();
+
+            return redirect(route("home")) ?: redirect(route("login_guest"));
         }
 
-        $user= User::where('id',Auth::user()->id)->first();
-        $user->session_id = Session::getId();
-        $user->save();
-
-        return redirect(route("home")) ?: redirect(route("login_guest"));
+        else{
+            Auth::logout();
+            $request->session()->invalidate();
+            $request->session()->regenerateToken();
+            return redirect(route("login_guest"));
+        }
     }
 }
